+++ neo/src/main/java/com/unrealdinnerbone/weathergate/attrubutes/Injector.java
@@ -14,16 +14,6 @@ public class Injector {
         for (TerrainModifier<?> terrainModifier : TerrainModifiers.MODIFIERS_REGISTRY) {
             environmentAttributes = addLayer(level, environmentAttributes, terrainModifier);
         }
-//        return environmentAttributes
-//                .addPositionalLayer(EnvironmentAttributes.SKY_LIGHT_COLOR, ofType(level, TerrainModifiers.SKY_LIGHT_COLOR))
-//                .addPositionalLayer(EnvironmentAttributes.FOG_COLOR, ofType(level, TerrainModifiers.FOG_COLOR))
-//                .addPositionalLayer(EnvironmentAttributes.WATER_FOG_COLOR, ofType(level, TerrainModifiers.WATER_FOG_COLOR))
-//                .addPositionalLayer(EnvironmentAttributes.SKY_COLOR, ofType(level, TerrainModifiers.SKY_COLOR))
-//                .addPositionalLayer(EnvironmentAttributes.SUNRISE_SUNSET_COLOR, ofType(level, TerrainModifiers.SUNRISE_SUNSET_COLOR))
-//                .addPositionalLayer(EnvironmentAttributes.CLOUD_COLOR, ofType(level, TerrainModifiers.CLOUD_COLOR))
-//                .addPositionalLayer(EnvironmentAttributes.BLOCK_LIGHT_TINT, ofType(level, TerrainModifiers.BLOCK_LIGHT_TINT))
-//                .addPositionalLayer(EnvironmentAttributes.NIGHT_VISION_COLOR, ofType(level, TerrainModifiers.NIGHT_VISION_COLOR))
-//                .addPositionalLayer(EnvironmentAttributes.AMBIENT_LIGHT_COLOR, ofType(level, TerrainModifiers.AMBIENT_LIGHT_COLOR));
         return environmentAttributes;
     }