+++ neo/src/main/java/com/unrealdinnerbone/weathergate/server/TerrainManager.java
@@ -1,6 +1,9 @@
 package com.unrealdinnerbone.weathergate.server;
 
 import com.unrealdinnerbone.weathergate.level.attachments.TerrainControllerAttachment;
+import com.unrealdinnerbone.weathergate.level.attachments.terrain.ControllerData;
+import com.unrealdinnerbone.weathergate.level.attachments.terrain.ModifierState;
+import com.unrealdinnerbone.weathergate.level.attachments.terrain.StoredData;
 import com.unrealdinnerbone.weathergate.registry.TerrainModifiers;
 import com.unrealdinnerbone.weathergate.modifers.base.TerrainModifier;
 import net.minecraft.core.BlockPos;
@@ -9,6 +12,7 @@ import net.minecraft.world.level.biome.Biome;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Optional;
 
 public class TerrainManager {
 
@@ -16,24 +20,24 @@ public class TerrainManager {
         TerrainControllerAttachment attachment = TerrainControllerAttachment.getAttachment(level);
         Biome biome = level.getBiome(pos).value();
 
-        Map<TerrainModifier<?>, TerrainControllerAttachment.ModifierState<?>> colorMap = new HashMap<>();
+        Map<TerrainModifier<?>, ModifierState<?>> colorMap = new HashMap<>();
         for (TerrainModifier<?> terrainModifier : TerrainModifiers.MODIFIERS_REGISTRY) {
             colorMap.put(
                     terrainModifier,
-                    new TerrainControllerAttachment.ModifierState<>(
+                    new ModifierState<>(
                             terrainModifier.isEnabledByDefault(),
                             terrainModifier.getDefaultValue(level, biome, pos)
                     )
             );
         }
 
-        attachment.data().put(pos, new TerrainControllerAttachment.StoredData(colorMap));
+        attachment.setData(pos, new ControllerData(ControllerData.DEFAULT_RANGE, new StoredData(colorMap)));
         attachment.save(level);
     }
 
     public static void removeBlockPos(Level level, BlockPos pos) {
         TerrainControllerAttachment attachment = TerrainControllerAttachment.getAttachment(level);
-        attachment.data().remove(pos);
+        attachment.removeData(pos);
         attachment.save(level);
     }
 }